security: fix PID validation, XBM integer overflow, NULL pointer dereferences, and cursor size parsing
- src/main.c: Replace atoi() with strtol() + validate pid > 0 to prevent
  kill(-1/0, signal) when LABWC_PID contains a non-positive or non-numeric
  value; add <errno.h> and <limits.h> headers

- src/img/img-xbm.c: Change tokenizer_context nr_tokens/alloc_tokens from
  'int' to 'size_t' to eliminate signed-integer overflow in the token array
  growth expression (alloc_tokens + 16) * 2

- src/config/session.c: Guard all three strdup_printf() return values before
  passing them to spawn_async_no_shell() which asserts command != NULL; a
  vsnprintf encoding error returns NULL and would otherwise crash

- src/input/cursor.c: Replace atoi(XCURSOR_SIZE) with strtoul() + explicit
  validation so that an invalid or zero env-var falls back to the default
  cursor size (24) instead of silently creating a 0-sized cursor manager;
  add <errno.h> and <stdlib.h> headers

Agent-Logs-Url: https://github.com/johanmalm/labwc/sessions/c4f45d9f-99b0-46fe-af24-56acf9d09fd7

+++ src/img/img-xbm.c
@@ -42,7 +42,7 @@ struct pixmap {
 struct tokenizer_context {
 	char *buffer_position;
 	struct token *tokens;
-	int nr_tokens, alloc_tokens;
+	size_t nr_tokens, alloc_tokens;
 };
 
 static void
